Move crypto1 to helpers, add it to the public API
