+++ lib/nfc/protocols/mf_classic/mf_classic_listener_i.h
@@ -3,7 +3,7 @@
 #include "mf_classic_listener.h"
 #include <lib/nfc/protocols/iso14443_3a/iso14443_3a_listener_i.h>
 #include <nfc/protocols/nfc_generic_event.h>
-#include "crypto1.h"
+#include <nfc/helpers/crypto1.h>
 
 #ifdef __cplusplus
 extern "C" {
